WIP: many much new classes
